+++ BambooTracker/gui/command/pattern/shrink_pattern_qt_command.cpp
@@ -9,12 +9,12 @@ ShrinkPatternQtCommand::ShrinkPatternQtCommand(PatternEditorPanel* panel, QUndoC
 
 void ShrinkPatternQtCommand::redo()
 {
-	panel_->update();
+	panel_->redrawByPatternChanged(true);
 }
 
 void ShrinkPatternQtCommand::undo()
 {
-	panel_->update();
+	panel_->redrawByPatternChanged(true);
 }
 
 int ShrinkPatternQtCommand::id() const